+++ src/itdelatrisu/opsu/beatmap/BeatmapSet.java
@@ -162,7 +162,7 @@ public class BeatmapSet implements Iterable<Beatmap> {
 				case "od": v = beatmap.overallDifficulty; break;
 				case "hp": v = beatmap.HPDrainRate; break;
 				case "bpm": v = beatmap.bpmMax; break;
-				case "length": v = beatmap.endTime / 1000; break;
+				case "length": v = (float) beatmap.endTime / 1000; break;
 				case "star":
 				case "stars": v = Math.round(beatmap.starRating * 100) / 100f; break;
 				default: return false;
